Rename get_protocol_constants.py to protocol_constants.py
