+++ concent_api/api-integration-force-report-computed-task-test.py
@@ -18,7 +18,7 @@ from utils.testing_helpers          import generate_ecc_key_pair
 
 from api_testing_helpers            import api_request
 
-from get_protocol_constants import get_protocol_constants
+from protocol_constants import get_protocol_constants
 
 import requests
